requirements.txt
customtkinter
pillow
tkinter
numpy